Move shared test fixtures out of src into tests/jest/fixtures
The enzyme suites under src/components/__tests__/ are gone, but the genre
tree and classifications data they shared stayed behind. Both are still
imported by four RTL suites, so move them next to those tests.

With src free of test code, the src/**/__tests__/** exclusions in
jest.config.js and .github/codecov.yml no longer match anything and are
dropped. Also drop a comment in Stats.test.tsx pointing at two migrated
suites, and refresh the README section describing the old mocha layout.
